replace flag for deepforest crown prediction

+++ src/data.py
@@ -326,43 +326,45 @@ class TreeData(LightningDataModule):
     def setup(self,stage=None):
         #Clean data from raw csv, regenerate from scratch or check for progress and complete
         if self.regenerate:
-            #remove any previous runs
-            try:
-                os.remove("{}/processed/canopy_points.shp".format(self.data_dir))
-                os.remove(" ".format(self.data_dir))
-                os.remove("{}/processed/crowns.shp".format(self.data_dir))
-                for x in glob.glob(self.config["crop_dir"]):
-                    os.remove(x)
-            except:
-                pass
+            if self.config["replace"]:#remove any previous runs
+                try:
+                    os.remove("{}/processed/canopy_points.shp".format(self.data_dir))
+                    os.remove(" ".format(self.data_dir))
+                    os.remove("{}/processed/crowns.shp".format(self.data_dir))
+                    for x in glob.glob(self.config["crop_dir"]):
+                        os.remove(x)
+                except:
+                    pass
+                    
+                #Convert raw neon data to x,y tree locatins
+                df = filter_data(self.csv_file, config=self.config)
                 
-            #Convert raw neon data to x,y tree locatins
-            df = filter_data(self.csv_file, config=self.config)
-            
-            #load any megaplot data
-            if not self.config["megaplot_dir"] is None:
-                megaplot_data = megaplot.load(directory=self.config["megaplot_dir"], rgb_pool=self.config["rgb_sensor_pool"], client = self.client, config=self.config)
-                #don't add species from contrib data (?) https://github.com/weecology/DeepTreeAttention/issues/65
-                megaplot_data = megaplot_data[megaplot_data.taxonID.isin(df.taxonID.unique())]
-                df = pd.concat([megaplot_data, df])
+                #load any megaplot data
+                if not self.config["megaplot_dir"] is None:
+                    megaplot_data = megaplot.load(directory=self.config["megaplot_dir"], rgb_pool=self.config["rgb_sensor_pool"], client = self.client, config=self.config)
+                    #don't add species from contrib data (?) https://github.com/weecology/DeepTreeAttention/issues/65
+                    megaplot_data = megaplot_data[megaplot_data.taxonID.isin(df.taxonID.unique())]
+                    df = pd.concat([megaplot_data, df])
+                    
+                #DEBUG, just one site
+                #df = df[df.siteID=="HARV"]
                 
-            #DEBUG, just one site
-            #df = df[df.siteID=="HARV"]
-            
-            #Filter points based on LiDAR height
-            df = CHM.filter_CHM(df, CHM_pool=self.config["CHM_pool"],min_CHM_diff=self.config["min_CHM_diff"], min_CHM_height=self.config["min_CHM_height"])      
-            df.to_file("{}/processed/canopy_points.shp".format(self.data_dir))
-            
-            #Create crown data
-            crowns = generate.points_to_crowns(
-                field_data="{}/processed/canopy_points.shp".format(self.data_dir),
-                rgb_dir=self.config["rgb_sensor_pool"],
-                savedir="{}/interim/".format(self.data_dir),
-                raw_box_savedir=None, 
-                client=self.client
-            )
-            
-            crowns.to_file("{}/processed/crowns.shp".format(self.data_dir))
+                #Filter points based on LiDAR height
+                df = CHM.filter_CHM(df, CHM_pool=self.config["CHM_pool"],min_CHM_diff=self.config["min_CHM_diff"], min_CHM_height=self.config["min_CHM_height"])      
+                df.to_file("{}/processed/canopy_points.shp".format(self.data_dir))
+                
+                #Create crown data
+                crowns = generate.points_to_crowns(
+                    field_data="{}/processed/canopy_points.shp".format(self.data_dir),
+                    rgb_dir=self.config["rgb_sensor_pool"],
+                    savedir="{}/interim/".format(self.data_dir),
+                    raw_box_savedir=None, 
+                    client=self.client
+                )
+                
+                crowns.to_file("{}/processed/crowns.shp".format(self.data_dir))
+            else:
+                crowns = gpd.read_file("{}/processed/crowns.shp".format(self.data_dir))
             
             annotations = generate.generate_crops(
                 crowns,
